+++ Sources/HTTP/Protocols/URLSession+Protocol.swift
@@ -2,11 +2,20 @@ import Foundation
 
 protocol URLSessionProtocol {
     func dataTask(with request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol
+    @available(iOS 15.0.0, *)
+    @available(macOS 12.0.0, *)
+    func data(for request: URLRequest) async throws -> (Data, URLResponse)
 }
 
 extension URLSession: URLSessionProtocol {
     func dataTask(with request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
         dataTask(with: request, completionHandler: completionHandler) as URLSessionDataTask
+    }
+    
+    @available(iOS 15.0.0, *)
+    @available(macOS 12.0.0, *)
+    func data(for request: URLRequest) async throws -> (Data, URLResponse) {
+        try await data(for: request, delegate: nil)
     }
 }