+++ Tests/HTTPTests/Doubles/URLSessionSpy.swift
@@ -19,4 +19,24 @@ final class URLSessionSpy: URLSessionProtocol {
         }
         return stubbedDataTaskResult
     }
+
+    var invokedData = false
+    var invokedDataCount = 0
+    var invokedDataParameters: (request: URLRequest, Void)?
+    var invokedDataParametersList = [(request: URLRequest, Void)]()
+    var stubbedDataError: Error?
+    var stubbedDataResult: (Data, URLResponse)!
+
+    @available(iOS 15.0.0, *)
+    @available(macOS 12.0.0, *)
+    func data(for request: URLRequest) async throws -> (Data, URLResponse) {
+        invokedData = true
+        invokedDataCount += 1
+        invokedDataParameters = (request, ())
+        invokedDataParametersList.append((request, ()))
+        if let error = stubbedDataError {
+            throw error
+        }
+        return stubbedDataResult
+    }
 }